pyproject.toml
[tool.poetry]
name = "pyinterval"
version = "0.1.0"
description = ""
authors = ["dcolon <djcolon90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydub = "^0.25.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pyInterval = "pyinterval.cli:main"